Render functions in their own page + doc.
