+++ src/game/visibility/view_systems.rs
@@ -1,137 +1,13 @@
 use std::cmp;
 use bresenham::Bresenham;
 
-use bevy::{prelude::*, utils::{HashMap, HashSet}};
+use bevy::{prelude::*, utils::HashSet};
 
-use crate::{engine::render::components::GameMapRender, game::{pieces::components::{Npc, Occupier}, player::Player, tileboard::components::{BoardPosition, Tile}}, map_builders::map::Map, vectors::Vector2Int};
+use crate::{ game::{pieces::components:: Occupier, player::Player, tileboard::components::{BoardPosition, Tile}}, map_builders::map::Map, vectors::Vector2Int};
 
 use super::components::{ChangeTileVisibility, ChangeTileVisibilityStatus, View};
 
-
-// 0.20c : Recupère les tuiles autour du personnage, en accord avec le range donné.
-// NOTE: Ne se préocupe pas des obstacles pour le moment.
-// BUG : Range de 10; Gauche et haut donnent bien -10 par rapport à ma position. Bas / Droite donnent seulement +9. TopLeft donne +10 les autres +9.
-// ==> OEUF CORSE. Range a...b : a inclusif, b exclusif...
-// Check with 0.20g => OK pour le Range respecté des deux cotés.
-fn get_tiles_around_range(
-    x: i32, 
-    y: i32,
-    range: i32,
-    max_x: i32, // map width
-    max_y: i32  // map height
- ) -> Vec<Vector2Int> {
-    let mut tiles_around_range : Vec<Vector2Int> = Vec::new();
-    for x in (cmp::max(x - range, 0))..(cmp::min(x + range, max_x) +1) {
-        for y in (cmp::max(y - range, 0))..(cmp::min(y + range, max_y) +1) {
-            tiles_around_range.push(Vector2Int {x, y} )
-        }
-    }
-    return tiles_around_range
- }
-
- // 0.20c Get Entity from game_map_render pour floor ou wall.
- fn get_floor_entity_at(
-    game_map_render:&GameMapRender,
-    x: i32,
-    y: i32
- ) -> Option<&Entity> {    
-    if game_map_render.floor.contains_key(&Vector2Int {x, y}) {
-        let option_entity_floor = game_map_render.floor.get(&Vector2Int {x, y});
-        option_entity_floor
-    } else {
-        None
-    }
- }
-
- fn get_wall_entity_at(
-    game_map_render:&GameMapRender,
-    x: i32,
-    y: i32
- ) -> Option<&Entity> {    
-    if game_map_render.wall.contains_key(&Vector2Int {x, y}) {
-        let option_entity_wall = game_map_render.wall.get(&Vector2Int {x, y});
-        option_entity_wall
-    } else {
-        None
-    }
- }
-
  
- // 0.20d mise à jour des tiles render.
- pub fn update_tile_visibility_render(
-    board: Res<Map>,
-    mut commands: Commands,
-    tile_with_change_order_q: Query<(Entity, &ChangeTileVisibility, &BoardPosition), With<Tile>>,
-    game_map_render_q: Query<&GameMapRender>, 
-    mut visibility_q: Query<&mut Visibility>,
- ){
-    let game_map_render = game_map_render_q.single();
-    let mut to_remove = Vec::new();
-    let mut tiles_to_change = HashMap::new();   // On place les infos dans ce Hashmap pour eviter les doublons de vector.
-
-    for (entity, new_visibility, position) in tile_with_change_order_q.iter() {
-        //info!("I have to render the following tiles : {:?}", position.v);
-        // Une tuile logique = 25% d'une tuile graphique x 4 dû au DualGrid. La position x,y corresponds à la partie Nord Ouest de la tuile logique.
-        // Il faut donc aussi traiter x-+1,y ; x, y+1 ; x+1,y+1. On ne doit pas depasser le board non plus.
-        // ===> En fait seul { x: cmp::min(position.v.x + 1, board.width - 1), y: position.v.y} permets d'avoir un affichage propre et ok avec le range. 
-        // Le reste semble s'overdriver et s'écraser mutuellement.
-        
-        //tiles_to_change.insert(Vector2Int { x: position.v.x, y: position.v.x }, new_visibility);
-        //tiles_to_change.insert(Vector2Int { x: cmp::min(position.v.x + 1, board.width - 1), y: cmp::min(position.v.y + 1, board.height - 1)}, new_visibility);
-        //tiles_to_change.insert(Vector2Int { x: position.v.x, y: cmp::min(position.v.y + 1, board.height - 1)}, new_visibility);
-        tiles_to_change.insert(Vector2Int { x: cmp::min(position.v.x + 1, board.width - 1), y: position.v.y}, new_visibility);
-
-        to_remove.push(entity);
-    }
-
-    for (tile_position, new_visibility) in tiles_to_change {
-        if let Some(entity_floor) = get_floor_entity_at(game_map_render, tile_position.x, tile_position.y ) {
-            if let Ok(mut visibility_floor) = visibility_q.get_mut(*entity_floor){
-                match new_visibility.new_status {
-                    ChangeTileVisibilityStatus::Visible => * visibility_floor = Visibility::Visible,
-                    ChangeTileVisibilityStatus::Hidden => * visibility_floor = Visibility::Hidden,
-                    ChangeTileVisibilityStatus::HiddenButKnown => {} //* visibility_floor = Visibility::Hidden,
-                }                
-            }                    
-        }
-        if let Some(entity_wall) = get_wall_entity_at(game_map_render, tile_position.x, tile_position.y ) {
-            if let Ok(mut visibility_wall) = visibility_q.get_mut(*entity_wall){
-                match new_visibility.new_status {
-                    ChangeTileVisibilityStatus::Visible => * visibility_wall = Visibility::Visible,
-                    ChangeTileVisibilityStatus::Hidden => * visibility_wall = Visibility::Hidden,
-                    ChangeTileVisibilityStatus::HiddenButKnown => {} //* visibility_floor = Visibility::Hidden,
-                }                
-            }                    
-        }
-    }       
-    for entity in to_remove {
-        commands.entity(entity).remove::<ChangeTileVisibility>();
-    }
- }
-
- // 0.20e ici on modifie l'affichage. L'intelligence "Je suis pas visible" va dans les autres systèmes.
- pub fn update_npc_visibility_status(
-    player_view_q: Query<&View, With<Player>>,
-    npc_position_q: Query<(Entity, &BoardPosition), With <Npc>>,
-    mut npc_visibility_q: Query<&mut Visibility, With<Npc>>,
- ){
-    for view in player_view_q.iter() {
-        let all_npc_positions:&HashSet<(Entity, Vector2Int)> = &npc_position_q.iter().map(|(npc_entity, npc_position)| (npc_entity, npc_position.v)).collect();
-        
-        //info!("My view is : {:?}", view.visible_tiles);
-        for (entity, position) in all_npc_positions{
-            let Ok(mut npc_visibility) = npc_visibility_q.get_mut(*entity) else { continue };
-            if view.visible_tiles.contains(position) {
-                //info!("Entity {:?} is in my view at {:?}", entity, position);                
-                *npc_visibility = Visibility::Visible;
-            } else {
-                //info!("Entity {:?} is not in view sight, because at {:?}", entity, position);
-                *npc_visibility = Visibility::Hidden;
-            }            
-        }
-    }
- }
-
  // 0.20f
  fn get_tiles_around_range_obstacles_break_view(  
     x: i32, 
@@ -206,28 +82,6 @@ fn get_tiles_around_range(
  }
 
 
- // 0.20f : Rework de get_tiles_around_range, n'affiche pas les obstacles. Pas utile, methode de debug.
- // Check with 0.20g => OK pour le Range respecté des deux cotés.
-fn get_tiles_around_range_with_obstacles(
-    x: i32, 
-    y: i32,
-    range: i32,
-    max_x: i32, // map width
-    max_y: i32,  // map height
-    obstacle_position_list: &HashSet< Vector2Int>
-
- ) -> Vec<Vector2Int> {
-    let mut tiles_around_range : Vec<Vector2Int> = Vec::new();
-    // Rappel : for x in a..b, b est exclusif.
-    for x in (cmp::max(x - range, 0))..(cmp::min(x + range, max_x) +1) {
-        for y in (cmp::max(y - range, 0))..(cmp::min(y + range, max_y) +1) {
-            if !obstacle_position_list.contains(&Vector2Int { x, y}) {
-                tiles_around_range.push(Vector2Int {x, y} )
-            }  
-        }
-    }
-    return tiles_around_range
- }
 
  // 0.20f
  pub fn update_character_view_with_blocked(
@@ -289,7 +143,60 @@ fn get_tiles_around_range_with_obstacles(
     }
  }
 
+
+
+// DEBUG : On affiche les tuiles autour, sans aucune contrainte.
+// 0.20c : Recupère les tuiles autour du personnage, en accord avec le range donné.
+// NOTE: Ne se préocupe pas des obstacles pour le moment.
+// BUG : Range de 10; Gauche et haut donnent bien -10 par rapport à ma position. Bas / Droite donnent seulement +9. TopLeft donne +10 les autres +9.
+// ==> OEUF CORSE. Range a...b : a inclusif, b exclusif...
+// Check with 0.20g => OK pour le Range respecté des deux cotés.
+fn get_tiles_around_range(
+    x: i32, 
+    y: i32,
+    range: i32,
+    max_x: i32, // map width
+    max_y: i32,  // map height,
+    obstacle_position_list: &HashSet< Vector2Int>       // Non utilisé.
+ ) -> Vec<Vector2Int> {
+    let mut tiles_around_range : Vec<Vector2Int> = Vec::new();
+    for x in (cmp::max(x - range, 0))..(cmp::min(x + range, max_x) +1) {
+        for y in (cmp::max(y - range, 0))..(cmp::min(y + range, max_y) +1) {
+            tiles_around_range.push(Vector2Int {x, y} )
+        }
+    }
+    return tiles_around_range
+ }
+
+
+ // DEBUG : On cache les obstacles de la vue, mais on ne prends pas en compte la ligne de vue : ce qui est derrière est affiché.
+ // 0.20f : Rework de get_tiles_around_range, n'affiche pas les obstacles. Pas utile, methode de debug.
+ // Check with 0.20g => OK pour le Range respecté des deux cotés.
+fn get_tiles_around_range_with_obstacles(
+    x: i32, 
+    y: i32,
+    range: i32,
+    max_x: i32, // map width
+    max_y: i32,  // map height
+    obstacle_position_list: &HashSet< Vector2Int>
+
+ ) -> Vec<Vector2Int> {
+    let mut tiles_around_range : Vec<Vector2Int> = Vec::new();
+    // Rappel : for x in a..b, b est exclusif.
+    for x in (cmp::max(x - range, 0))..(cmp::min(x + range, max_x) +1) {
+        for y in (cmp::max(y - range, 0))..(cmp::min(y + range, max_y) +1) {
+            if !obstacle_position_list.contains(&Vector2Int { x, y}) {
+                tiles_around_range.push(Vector2Int {x, y} )
+            }  
+        }
+    }
+    return tiles_around_range
+ }
+
+
+ // Original version. Remplacée par update_character_view_with_blocked
  // 0.20d visibility system with component. Only works for Logic Tile.     
+ /* 
 pub fn update_character_view(
     mut commands: Commands,
     mut player_view_q: Query<(&mut View, &BoardPosition), With<Player>>,
@@ -351,3 +258,4 @@ pub fn update_character_view(
         view.visible_tiles = current_view;
     }
 }
+*/
